+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### July 10, 2024
 1. Implemented HSV color picker on DiaryEntry.kt
 2. Modified /self to include TODO list
+3. Modified overall app theme and typography
+4. Refactored PresetParams.kt to Reusables.kt
+5. Implemented second iteration of DiaryEntryPage.kt user interface
 
 
 ### July 6, 2024
@@ -34,6 +37,4 @@
 6. Implemented Navigation Compose libraries.
 
 # TODO:
-1. Overhaul DiaryEntry.kt UI
-2. Refactor PresetParams.kt -> Reusables.kt
-3. Implement business logic on DiaryEntry.kt
+1. Implement business logic on DiaryEntry.kt